Move db-cache into separate file
